+++ src/main/java/org/mule/module/hubspot/model/token/RefreshTokenRequest.java
@@ -1,0 +1,58 @@
+/**
+ *
+ * (c) 2003-2012 MuleSoft, Inc. This software is protected under international
+ * copyright law. All use of this software is subject to MuleSoft's Master
+ * Subscription Agreement (or other Terms of Service) separately entered
+ * into between you and MuleSoft. If such an agreement is not in
+ * place, you may not use the software.
+ */
+
+package org.mule.module.hubspot.model.token;
+
+import org.codehaus.jackson.annotate.JsonIgnoreProperties;
+import org.codehaus.jackson.annotate.JsonProperty;
+import org.codehaus.jackson.map.annotate.JsonSerialize;
+import org.codehaus.jackson.map.annotate.JsonSerialize.Inclusion;
+
+@JsonSerialize(include = Inclusion.NON_NULL)
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class RefreshTokenRequest {
+	
+	private String refreshToken;
+	private String clientId;
+	private String grantType;
+	
+	@JsonProperty("refresh_token")
+	public String getRefreshToken() {
+		return refreshToken;
+	}
+	
+	@JsonProperty("refresh_token")
+	public void setRefreshToken(String refreshToken) {
+		this.refreshToken = refreshToken;
+	}
+
+	@JsonProperty("client_id")
+	public String getClientId() {
+		return clientId;
+	}
+
+	@JsonProperty("client_id")
+	public void setClientId(String clientId) {
+		this.clientId = clientId;
+	}
+
+	@JsonProperty("grant_type")
+	public String getGrantType() {
+		return grantType;
+	}
+
+	@JsonProperty("grant_type")
+	public void setGrantType(String grantType) {
+		this.grantType = grantType;
+	}
+	
+	public String toString() {
+		return String.format("refresh_token=%s&client_id=%s&grant_type=%s", refreshToken, clientId, grantType);
+	}
+}